requirement.txt
fastapi[standard]
google-generativeai
PyPDF2
python-dotenv
python-multipart
uvicorn
reportlab
